Light remains on until object not OIFOV
Now the device will appear more responsive, as it will LED indicate when
OIFOV only instead of a fixed on/off time that will happen each time
(i.e. before the device would blink if a device was in the field, now it
will be constant until it leaves).   Need to decide if we want this
change or not and need to validate with testing.