Correct dynamic OIFOV, moved low battery hdl
Added in hardware tested code to just live detect the OIFOV instead of
the fixed time version.  Feels more responsive this way.  Also removed
the low battery handling outside all the data/sleep if statements.  Will
allow for easier modifications as this is still not finalized.  Will
probably do some sort of flashing to allow the device to charge up as a
constant low battery LED on will guarantee the battery will die and user
must take it off to charge.  Also changed the sonar sample rate to be
100ms to allow enough time in between.  This can be played with if
needed: the longer between the better as the sonar is off during this
time (the off on time may be too slow and we may need to just leave on
the USS, this still needs further testing).