PNP workaround for USS pwr, turning off 5V output
The transistor setup was not working, so shutdown the 5V output when
going into deep sleep, enable it when coming out of deep sleep (from
motion).  Transistor needs to be NPN from USS GND pin to digital GND,
then the transistor will properly work.  For now swapped in a PNP
resistor so switch all SONIC_PWR_PORT to be active low instead of high.
Will need to change if going back to an NPN transistor.